requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.2.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
configparser==5.2.0
contourpy==1.0.6
crayons==0.4.0
cryptography==36.0.0
cycler==0.11.0
defusedxml==0.7.1
distlib==0.3.4
django-allauth==0.47.0
django-environ==0.8.1
djangorestframework==3.13.1
filelock==3.6.0
fonttools==4.38.0
fpdf==1.7.2
fpdf2==2.5.0
frozenlist==1.3.0
geoip2==4.5.0
h11==0.12.0
idna==3.3
kiwisolver==1.4.4
matplotlib==3.6.2
maxminddb==2.2.0
multidict==6.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.22.1
oauthlib==3.1.1
outcome==1.1.0
packaging==21.3
pandas==1.4.0
Pillow==9.0.0
platformdirs==2.5.2
protobuf==3.19.4
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.3
razorpay==1.3.0
requests==2.26.0
requests-oauthlib==1.3.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
virtualenv==20.14.1
webdriver-manager==3.5.2
wsproto==1.0.0
xlrd==2.0.1
yarl==1.7.2
